+++ lib/index.js
@@ -5,30 +5,8 @@ import { CompositeDisposable } from 'atom';
 import path from 'path';
 
 import handleData from './message.js';
-import { check } from './helpers';
-
-import type { LinterMessage } from './message';
-
-
-type BufferType = {
-  cachedText: string;
-  getText(): string;
-}
-
-type TextEditorType = {
-  getPath(): string;
-  getText(): string;
-  buffer: BufferType;
-  isModified(): boolean;
-}
-
-type Linter = {
-  grammarScopes: Array<string>;
-  scope: 'file' | 'project';
-  name?: string;
-  lintOnFly: boolean;
-  lint(TextEditor: TextEditorType): Array<LinterMessage> | Promise<Array<LinterMessage>>;
-};
+import type { Linter } from './types.js';
+import { check } from './helpers.js';
 
 export default {
   config: {
