+++ lib/message.js
@@ -3,32 +3,8 @@
 
 import { Range } from 'atom';
 
-export type LinterTrace = {
-  type: 'Trace';
-  text?: string;
-  html?: string;
-  filePath: string;
-  range?: Range;
-};
+import type { FlowError, LinterTrace, LinterMessage } from './types.js';
 
-export type LinterMessage = {
-  type: 'Error' | 'Warning';
-  text?: string;
-  html?: string;
-  filePath?: string;
-  range?: Range;
-  trace?: Array<LinterTrace>;
-};
-
-export type FlowError = {
-  level: string;
-  descr: string;
-  path: string;
-  line: number;
-  start: number;
-  endline: number;
-  end: number;
-};
 
 function extractRange(message: FlowError): Range {
   return new Range(
